back-end/services/app/requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1665754496418/work
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1666878637568/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1659291887007/work
bcrypt==4.0.1
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1664823096650/work
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764759924/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachelib==0.9.0
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1660055382950/work
certifi==2022.12.7
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1671179893800/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1669592251784/work
DateTime==4.8
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
flask-redis==0.4.0
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1670726765190/work
frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1667935502123/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1596839682936/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1669953359096/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpc-split_1666813479356/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1666831184106/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1669312071043/work
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1662254869273/work/keras-2.10.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1657921053229/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1666770346139/work
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1666830309128/work
mysql-connector-python==8.0.31
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1668919228101/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1666056362788/work
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1617859230218/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1670530880680/work
pandas==1.5.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyJWT==2.0.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1665350324128/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
pytz==2022.6
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
redis==4.4.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1658328885051/work
scipy==1.9.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.38.3
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1664238338171/work/tensorboard-2.10.1-py3-none-any.whl
tensorboard-data-server @ file:///Users/runner/miniforge3/conda-bld/tensorboard-data-server_1670043875829/work/tensorboard_data_server-0.6.1-py3-none-macosx_11_0_arm64.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow @ file:///Users/ngam/Repos/tensorflow-feedstock/miniforge3/conda-bld/tensorflow-split_1663963015006/work/tensorflow_pkg/tensorflow-2.10.0-cp310-cp310-macosx_11_0_arm64.whl
tensorflow-estimator @ file:///Users/ngam/Repos/tensorflow-feedstock/miniforge3/conda-bld/tensorflow-split_1663963015006/work/tensorflow-estimator/wheel_dir/tensorflow_estimator-2.10.0-py2.py3-none-any.whl
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1669015807632/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1669259737463/work
uuid==1.30
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1660053523487/work
wrapt @ file:///Users/runner/miniforge3/conda-bld/wrapt_1666806247108/work
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1666861803792/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1669453021653/work
zope.interface==5.5.2
python-dotenv
flask-cors